script now allows you to alternate between running and stopped without reselecting the text, behavior like Mac OSX

+++ tts.sh
@@ -1,3 +1,10 @@
 #!/bin/bash
-killall espeak # this stops (interrupts) the currently running thread if there is one 
-xsel -c | espeak --stdin # xsel -c clears the value after returning it, the return is piped into espeak via standerd in
+# this detects weather espeak is running 
+if pidof espeak > 0
+then
+	# this stops (interrupts) the currently running thread if there is one
+	killall espeak
+else
+	# xsel returns the slected text and pipes it into espeak standerd in
+	xsel | espeak --stdin
+fi
